오후 5:42

+++ FinalProject/src/main/java/kr/co/project/common/upload/MultiUploadFile.java
@@ -21,8 +21,7 @@ import kr.co.project.recipe.model.dto.RecipeDTO;
 
 public class MultiUploadFile {
 
-	private static final String UPLOAD_PATH="C:\\Users\\pje97\\git\\finalProject\\FinalProject\\src\\main\\webapp\\resources\\uploads\\";
-
+	private static final String UPLOAD_PATH = "C:\\Users\\rlwjd\\git\\finalProject\\FinalProject\\src\\main\\webapp\\resources\\uploads\\";
 
 	public static boolean deleteFile(List<RecipeDTO> fileName, List<RecipeDTO> sequenceFileName) {
 		// C:\\spring\\Project\\src\\main\\webapp\\resources\\uploads\\
@@ -98,53 +97,55 @@ public class MultiUploadFile {
 		return "result";
 	}
 
-	public static String goodsUploadMethod(
-			List<MultipartFile> multiFileList, 
-			HttpSession session,
-			GoodsDTO goods,
-			List<GoodsPhotosDTO> goodsPhotoList) {
-
+	public static String goodsUploadMethod(List<MultipartFile> multiFileList, HttpSession session, GoodsDTO goods,
+			List<GoodsPhotosDTO> goodsPhotoList, List<GoodsPhotosDTO> detailPhotoList) {
 		List<Map<String, String>> fileList = new ArrayList<>();
-
 		for (int i = 0; i < multiFileList.size(); i++) {
-			String originFile = multiFileList.get(i).getOriginalFilename();//사진 원래 이름
-			System.out.println(originFile);
-			String ext = originFile.substring(originFile.lastIndexOf(".")); //확장자
-			String changeFile = UUID.randomUUID().toString() + ext; //이름 숫자열덤 +확장자
-			// 데이터포맷을 년월일시분초로 가공
-
+			String originFile = multiFileList.get(i).getOriginalFilename();// 사진 원래 이름
 			Map<String, String> map = new HashMap<>();
-			map.put("originFile", originFile);
-			map.put("changeFile", changeFile);
-
-			GoodsPhotosDTO gps = new GoodsPhotosDTO();
-
-			if (i == 0) {//1st 사진은 goods에 저장
-				goods.setG_origin_name(originFile);
-				goods.setG_file_name(changeFile);
-				goods.setG_file_path(UPLOAD_PATH + "goods\\" + changeFile);
-			} else {//이후로는 goodsphoto 저장
-				gps.setGp_origin_name(originFile);
-				gps.setGp_file_name(changeFile);
-				gps.setGp_file_path(UPLOAD_PATH + "goods\\" + changeFile);
+			if (originFile.length() >= 1) {
+				System.out.println(i + "번째 사진" + originFile + " 글자수 :" + originFile.length());
+				String ext = originFile.substring(originFile.lastIndexOf(".")); // 확장자
+				String changeFile = UUID.randomUUID().toString() + ext; // 이름 숫자열덤 +확장자
+				// 데이터포맷을 년월일시분초로 가공
+				map.put("originFile", originFile);
+				map.put("changeFile", changeFile);
+				GoodsPhotosDTO gps = new GoodsPhotosDTO();
+				if (i == 0) {
+					// 1st 사진은 goods에 저장
+					goods.setG_origin_name(originFile);
+					goods.setG_file_name(changeFile);
+					goods.setG_file_path(UPLOAD_PATH + "goods\\" + changeFile);
+				} else if (i > 0 && i <= 4) {
+					// 2~5 goodsphoto 저장
+					gps.setGp_origin_name(originFile);
+					gps.setGp_file_name(changeFile);
+					gps.setGp_file_path(UPLOAD_PATH + "goods\\" + changeFile);
+					goodsPhotoList.add(gps); // 리스트
+				} else if (i > 4 && i <= 9) {
+					gps.setGp_origin_name(originFile);
+					gps.setGp_file_name(changeFile);
+					gps.setGp_file_path(UPLOAD_PATH + "goods\\" + changeFile);
+					detailPhotoList.add(gps); // 리스트
+				}
 			}
-
-			goodsPhotoList.add(gps);//리스트
 			fileList.add(map);
 		}
-
 		try {
 			for (int i = 0; i < multiFileList.size(); i++) {
-				File uploadFile = new File(UPLOAD_PATH + "goods\\" + fileList.get(i).get("changeFile"));
-				multiFileList.get(i).transferTo(uploadFile);
+				if (!fileList.get(i).isEmpty()) {
+					File uploadFile = new File(UPLOAD_PATH + "goods\\" + fileList.get(i).get("changeFile"));
+					multiFileList.get(i).transferTo(uploadFile);
+				}
 			}
 		} catch (IllegalStateException | IOException e) {
 			for (int i = 0; i < multiFileList.size(); i++) {
-				new File(UPLOAD_PATH + "goods\\" + fileList.get(i).get("changeFile")).delete();
+				if (!fileList.get(i).isEmpty()) {
+					new File(UPLOAD_PATH + "goods\\" + fileList.get(i).get("changeFile")).delete();
+				}
 			}
 			e.printStackTrace();
 		}
 		return "result";
 	}
-
 }
